rag/requirements.txt
fastapi==0.98.0
uvicorn==0.23.2
PyPDF2==3.0.1
google-generativeai==0.1.0
sentence-transformers==2.2.2
scikit-learn==1.3.0
numpy==1.24.3
python-multipart==0.0.6
